extract to component, add download button
