+++ package.json
@@ -11,11 +11,13 @@
   },
   "dependencies": {
     "@now/node": "^1.4.0",
+    "@types/downloadjs": "^1.4.1",
     "@types/express": "^4.17.2",
     "@types/flat": "0.0.28",
     "@types/lodash": "^4.14.149",
     "@types/yaml": "^1.2.0",
     "axios": "^0.19.2",
+    "downloadjs": "^1.4.7",
     "express": "^4.17.1",
     "flat": "^5.0.0",
     "lodash": "^4.17.15",
