+++ src/views/Test.vue
@@ -1,139 +1,27 @@
 <template>
   <div>
     <h3>Testing: {{row.function_name}}</h3>
-    <ul>
-      <li v-for="(message, i) in messages" :key="i" class="message">{{message}}</li>
-    </ul>
 
-    <div class="tabs two">
-      <input ref="tab1" id="tab-1" type="radio" name="tabgroupB" checked />
-      <label class="pseudo button toggle" for="tab-1">Request</label>
-      <input ref="tab2" id="tab-2" type="radio" name="tabgroupB" />
-      <label class="pseudo button toggle" for="tab-2">Response</label>
-
-      <div class="row">
-        <!-- REQUEST -->
-        <div>
-          <textarea
-            v-model="request"
-            rows="20"
-            @input="update_validity"
-            :disabled="sending_request"
-          ></textarea>
-
-          <footer v-if="!sending_request">
-            <button class="success" :disabled="!request_valid" @click="test">Go!</button>
-            <span class="notify" v-if="!request_valid && request">Input is not valid JSON</span>
-          </footer>
-        </div>
-
-        <!-- RESPONSE -->
-        <div>
-          <div v-if="response">
-            <pre>{{formattedResponse}}</pre>
-          </div>
-          <div v-else-if="!sending_request">Waiting for request to be sent.</div>
-          <div v-else>Waiting for request to complete...</div>
-        </div>
-      </div>
-    </div>
-    <button @click="$router.go(-1)">Back to list</button>
+    <Requester :row="row" :fn="test" :get_sample="get_test_req_json" />
   </div>
 </template>
 
 <script lang="ts">
 import Vue from 'vue';
-import { OutgoingCombinedRecord, BusActions } from '@/types';
 import { get_test_req_json, test } from '@/test';
-import router from '@/router';
-import { EventBus } from '../event_bus';
+import { OutgoingCombinedRecord } from '@/types';
+import Requester from '@/components/Requester.vue';
+
 export default Vue.extend({
-  name: 'test',
+  components: { Requester },
   data() {
     return {
-      sending_request: false,
-      request: null as null | any,
-      request_valid: false,
-      response: null as null | string,
-      messages: [] as string[],
+      test,
+      get_test_req_json,
     };
   },
   props: {
     row: Object as () => OutgoingCombinedRecord,
   },
-  computed: {
-    formattedResponse(): string {
-      if (this.response === null) {
-        return '';
-      }
-      try {
-        return JSON.stringify(JSON.parse(this.response), null, 2);
-      } catch {
-        return this.response;
-      }
-    },
-  },
-  mounted() {
-    this.try_get_test_json();
-  },
-  methods: {
-    update_validity() {
-      this.request_valid = this.check_json_validity(this.request);
-    },
-    async try_get_test_json() {
-      this.messages.push(`Fetching test_req.json for ${this.row.function_name}`);
-      try {
-        const value = await get_test_req_json(this.row);
-        if (this.check_json_validity(value)) {
-          this.messages.push(`Successfully fetched test req from repo`);
-          this.request = JSON.stringify(JSON.parse(value), null, 2);
-        } else {
-          this.messages.push(`Test req could not be found, check if repo exists and a 'test_req.json' file is in the root folder`);
-          this.request = JSON.stringify({});
-        }
-        this.update_validity();
-      } catch (error) {
-        throw error;
-      }
-    },
-    async test() {
-      this.sending_request = true;
-      this.messages.push(`Sending request...`);
-
-      try {
-        (this.$refs.tab2 as HTMLInputElement).click();
-        const start = Date.now();
-        const value = await test(this.row.function_name, JSON.parse(this.request));
-        const end = Date.now();
-
-        this.response = value;
-        this.messages.push(`Results of running ${this.row.function_name}`);
-        this.messages.push(`Runtime : ${(end - start) / 1000} seconds`);
-      } catch (error) {
-        throw error;
-      }
-    },
-    check_json_validity(json: any): boolean {
-      try {
-        const obj = JSON.parse(json);
-        return (obj && typeof obj === 'object' && obj !== null);
-      } catch (err) {
-        return false;
-      }
-    },
-  },
 });
 </script>
-
-<style lang="css" scoped>
-.notify {
-  color: red;
-  margin-left: 10px;
-}
-.message {
-  color: grey;
-}
-.message:last-child {
-  color: inherit;
-}
-</style>
